Move tests into parser
