+++ src/arg_parsers/mod.rs
@@ -14,4 +14,4 @@ pub use filename::*;
 mod numbers;
 pub use numbers::*;
 mod identify_format;
-pub use identify_format::*;
+pub use identify_format::{IdentifyFormat, Token, Var};
